+++ CHANGELOG.md
@@ -1,6 +1,13 @@
+# 0.12.0
+
+- DONE: Add the whole navigator
+- ADD: `NAVS`in `@screens` to have constantes for routes
+- FIX: Navigation in `ScreenCar`and `ScreenDriver`
+- 
+
 # 0.11.0
 
-- DONE: Link hte new modal of pictures to the waypoint dashboard
+- DONE: Link the new modal of pictures to the waypoint dashboard
 - FIX: `name`and `address`property for `ScreenWaypointGalery`
 - ADD: A new look for disabled `CButton`
 - ADD: The photo button in `CWaypointButtons`now contain the number of pictures and is disabled if no pictures